Reset to new Figma plugin boilerplate
