+++ README.md
@@ -1,58 +1,19 @@
-Below are the steps to get your plugin running. You can also find instructions at:
+# Mentarii Figma Plugin
 
-  https://www.figma.com/plugin-docs/plugin-quickstart-guide/
+This is a simple boilerplate Figma plugin built with TypeScript and React. The plugin demonstrates how to show a UI built with React and communicate with the Figma plugin environment.
 
-This plugin template uses Typescript and NPM, two standard tools in creating JavaScript applications. It also relies on React and Vite for the plugin UI.
+## Development
 
-First, download Node.js which comes with NPM. This will allow you to install TypeScript and other
-libraries. You can find the download link here:
+1. Install dependencies
+   ```
+   npm install
+   ```
 
-  https://nodejs.org/en/download/
+2. Build the plugin
+   ```
+   npm run build
+   ```
 
-Next, install TypeScript using the command:
+3. In Figma, select **Plugins > Development > Import plugin from manifest...** and choose the `manifest.json` file in this directory.
 
-  npm install -g typescript
-
-Finally, in the directory of your plugin, get the latest type definitions for the plugin API by running:
-
-  npm install --save-dev @figma/plugin-typings
-
-If you are familiar with JavaScript, TypeScript will look very familiar. In fact, valid JavaScript code
-is already valid Typescript code.
-
-TypeScript adds type annotations to variables. This allows code editors such as Visual Studio Code
-to provide information about the Figma API while you are writing code, as well as help catch bugs
-you previously didn't notice.
-
-For more information, visit https://www.typescriptlang.org/
-
-Using TypeScript requires a compiler to convert TypeScript (code.ts) into JavaScript (code.js)
-for the browser to run.
-
-We recommend writing TypeScript code using Visual Studio code:
-
-1. Download Visual Studio Code if you haven't already: https://code.visualstudio.com/.
-2. Open this directory in Visual Studio Code.
-3. Compile TypeScript to JavaScript: Run the "Terminal > Run Build Task..." menu item,
-    then select "npm: watch". You will have to do this again every time
-    you reopen Visual Studio Code.
-
-That's it! Visual Studio Code will regenerate the JavaScript file every time you save.
-
-### Building the plugin UI with Vite
-
-If you are using React and Vite for the UI, run the build command before testing the plugin in Figma:
-
-```
-npm install
-npm run build
-```
-
-The `npm run build` command compiles `code.ts` and bundles the React UI into the `dist/` folder. This step generates `code.js` and `dist/index.html`, which Figma reads as specified in `manifest.json`. If these files don't exist, the React UI will not appear and Figma shows "an error occurred while loading the plugin environment" when you try to run the plugin.
-
-Make sure your `manifest.json` points to the built files, for example:
-
-```
-"main": "code.js",
-"ui": "dist/index.html"
-```
+See the [Figma plugin documentation](https://www.figma.com/plugin-docs/) for more details.
