+++ manifest.json
@@ -1,16 +1,7 @@
 {
-  "name": "Mentarii",
-  "id": "1527180713478689544",
+  "name": "Mentarii Figma Plugin",
   "api": "1.0.0",
-  "main": "code.js",
-  "capabilities": [],
-  "enableProposedApi": false,
-  "documentAccess": "dynamic-page",
-  "editorType": ["figma"],
+  "main": "dist/code.js",
   "ui": "dist/index.html",
-  "networkAccess": {
-    "allowedDomains": [
-      "none"
-    ]
-  }
+  "id": "com.example.mentarii"
 }
