--- ui/App.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function App() {
-  return <div>Hello from React + Vite in Figma!</div>;
-}
