--- ui/main.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { createRoot } from 'react-dom/client';
-import App from './App';
-
-const container = document.getElementById('root');
-if (container) {
-  createRoot(container).render(<App />);
-} else {
-  throw new Error("Root element not found");
-}
